Fix bug with data.sql trying to insert into ShopProduct table
When trying to perform INSERT INTO ShopProducts with @Table(name = "ShopProducts") on the class, Spring Boot would be unable to load the application context, saying that it could not find the ShopProducts table. Turns out this happens because Hibernate's default NamingStrategy automatically converts camel case and pascal case into snake case. Changing the sql to Shop_Products fixed the issue.